+++ spimdisasm/mips/sections/MipsSectionBss.py
@@ -13,7 +13,7 @@ from . import SectionBase
 
 
 class SectionBss(SectionBase):
-    def __init__(self, context: common.Context, vromStart: int, vromEnd: int, bssVramStart: int, bssVramEnd: int, filename: str, segmentVromStart: int, overlayCategory: str|None):
+    def __init__(self, context: common.Context, vromStart: int, vromEnd: int, bssVramStart: int, bssVramEnd: int, filename: str, segmentVromStart: int, overlayCategory: str|None) -> None:
         super().__init__(context, vromStart, vromEnd, bssVramStart, filename, [], common.FileSectionType.Bss, segmentVromStart, overlayCategory)
 
         self.bssVramStart: int = bssVramStart
@@ -29,13 +29,13 @@ class SectionBss(SectionBase):
     def sizew(self) -> int:
         return self.bssTotalSize // 4
 
-    def setVram(self, vram: int):
+    def setVram(self, vram: int) -> None:
         super().setVram(vram)
 
         self.bssVramStart = vram
         self.bssVramEnd = vram + self.bssTotalSize
 
-    def analyze(self):
+    def analyze(self) -> None:
         self.checkAndCreateFirstSymbol()
 
         # If something that could be a pointer found in data happens to be in the middle of this bss file's addresses space