+++ spimdisasm/mips/sections/MipsSectionRodata.py
@@ -15,7 +15,7 @@ from . import SectionBase
 
 
 class SectionRodata(SectionBase):
-    def __init__(self, context: common.Context, vromStart: int, vromEnd: int, vram: int, filename: str, array_of_bytes: bytes, segmentVromStart: int, overlayCategory: str|None):
+    def __init__(self, context: common.Context, vromStart: int, vromEnd: int, vram: int, filename: str, array_of_bytes: bytes, segmentVromStart: int, overlayCategory: str|None) -> None:
         if common.GlobalConfig.ENDIAN_RODATA is not None:
             words = common.Utils.endianessBytesToWords(common.GlobalConfig.ENDIAN_RODATA, array_of_bytes, vromStart, vromEnd)
         else:
@@ -25,7 +25,7 @@ class SectionRodata(SectionBase):
         self.stringEncoding = common.GlobalConfig.RODATA_STRING_ENCODING
 
 
-    def analyze(self):
+    def analyze(self) -> None:
         self.checkAndCreateFirstSymbol()
 
         symbolList = []