+++ spimdisasm/mips/sections/MipsSectionText.py
@@ -16,7 +16,7 @@ from . import SectionBase
 
 
 class SectionText(SectionBase):
-    def __init__(self, context: common.Context, vromStart: int, vromEnd: int, vram: int, filename: str, array_of_bytes: bytes, segmentVromStart: int, overlayCategory: str|None):
+    def __init__(self, context: common.Context, vromStart: int, vromEnd: int, vram: int, filename: str, array_of_bytes: bytes, segmentVromStart: int, overlayCategory: str|None) -> None:
         super().__init__(context, vromStart, vromEnd, vram, filename, common.Utils.bytesToWords(array_of_bytes, vromStart, vromEnd), common.FileSectionType.Text, segmentVromStart, overlayCategory)
 
         self.instrCat: rabbitizer.Enum = rabbitizer.InstrCategory.CPU
@@ -277,7 +277,7 @@ class SectionText(SectionBase):
         return funcsStartsList, unimplementedInstructionsFuncList
 
 
-    def analyze(self):
+    def analyze(self) -> None:
         instrsList = self.wordListToInstructions(self.words, self.getVramOffset(0), self.instrCat)
         nInstr = len(instrsList)
 
@@ -334,7 +334,7 @@ class SectionText(SectionBase):
         self.fileBoundaries = sorted(set(self.fileBoundaries))
 
 
-    def compareToFile(self, other: FileBase):
+    def compareToFile(self, other: FileBase) -> dict:
         result = super().compareToFile(other)
 
         if isinstance(other, SectionText):